requirements.txt
requests
python-dotenv
SQLAlchemy
pandas
numpy
matplotlib